+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Added
+- `action_source` field to EventRequest for Conversions API.
+
 ## v9.0.0
 
 
@@ -119,4 +122,3 @@ All notable changes to this project will be documented in this file.
     $async_job = $async_job->getSelf();
   }
   ```
-