+++ CHANGELOG.md
@@ -5,9 +5,6 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
-### Added
-- `action_source` field to EventRequest for Conversions API.
-
 ## v9.0.0
 
 
@@ -122,3 +119,4 @@ All notable changes to this project will be documented in this file.
     $async_job = $async_job->getSelf();
   }
   ```
+